Improve makefiles
Move shared makefile code in to lib/simplified.mk. This can be deployed
as a separate project for general reuse. Users can write a simple
makefile the includes simplified.mk and just specify the files to be
included in the build. Simplified.mk takes care of building the library
or executable; provides clean, tests and install targets.